+++ packages/core/CHANGELOG.md
@@ -1,4 +1,4 @@
-## Next
+## 1.0.0-beta.34
 
 * Bug fix: actually pass arguments object to `on.initial()`.